requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.23.5
matplotlib>=3.7.1
pandas>=2.1.0
scikit-learn>=1.2.2
Pillow>=9.5.0
